Partial fix for Linux
Still if needed to identify why the Qt libraries are not limited to have
the major version in the name, possible is because author changed it
before publishing unaware of the consequences of his actions. By the
other part because I have the Qt used in the project in my sytem and I
don't want to waste to much time in getting, and somthing similar with
libicu 58.0, the filename of the cmake_current_source_files may still
need correction.