Linux Partial Fix Corrected
The previous "patch" Linux Partial Fix, incorrectly tried to link the
ICU libs to the main binary, when they only need to be linked by the Qt
libs, and as the test mentioned in
https://github.com/Trico-Everfire/p2-publishing-tool/issues/6 shows that
that the Qt binaries autodetect the ICU in local path. Thus only need to
copy the ICU binaries. The naming problem of the ICU binaries was solved
just giving the instruction of use of certain name in the new directory lib/ICU/Readme.txt a copy of libs/Qt/Readme.txt with the appropiate changes.

Note that still is a mystery why the bundled Qt binaries are named until
the patch when the copy commend in the CMakeLists.txt clearly states
that they should be named until the major. And the linkind requier it
so.

Finally is important to note that shiping more libraries may be still be
needed to ensure that it can be runned in all kernel compatible linux
system, with GNU standard c libray, and X11 window system.